stage 1 covariance matrices are similar when normalized
compared my cov1 to og matlab cov1, where each was divided by its max
element so all values are from 0 to 1. Max difference was 1e-2
Max difference when compares to randn() was 1.5, so I think I got
covariance of patches right. Not sure why the scale is different. Might
be because of batch size